Improves the `README.md` and merge it with the LICENSE statement.
The license used by this library is a modified version of the
Modelica License 2, this was not clear when pointing to the original
modelica.org hosted version. Therefore those references were removed.
That little text included in README.md. It might be a good idea to
recreate `LICENCE.md` but this time containing the full licence text as
written in `OpenHydraulics.UsersGuide.ModelicaLicence2`